refactor, always create a new ModalInstance when opening
