+++ src/components/Backdrop.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import useSheetModal from "./hooks/useSheetModal";
+import useSheetModal from "../hooks/useSheetModal";
 import Animated from "react-native-reanimated";
-import useStableAnimatedStyle from "./hooks/useStableAnimatedStyle";
+import useStableAnimatedStyle from "../hooks/useStableAnimatedStyle";
 
 function SheetModalBackdrop() {
   const store = useSheetModal();
